+++ foxclaw/nodes/__init__.py
@@ -1,0 +1,19 @@
+"""FoxClaw node coordination helpers."""
+
+from .apollo import (
+    ApolloNodeBrief,
+    GitSnapshot,
+    NodeAuthorityLocks,
+    build_node_brief,
+    git_snapshot,
+    render_markdown,
+)
+
+__all__ = [
+    "ApolloNodeBrief",
+    "GitSnapshot",
+    "NodeAuthorityLocks",
+    "build_node_brief",
+    "git_snapshot",
+    "render_markdown",
+]